refactor(noti): decoupling `makeNoti` function
